+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 ## v0.4.25
 
 FEATURE: Public share oauth email domains renamed to `email address patterns`. These now take a glob to match against.
-Old `zrok share` and `zrok reserve` flag: `oauth-email-domains`. Changed to: `oauth-email-address-patterns`.
+Old `zrok share` and `zrok reserve` flag: `oauth-email-domains`.
+Changed to: `oauth-email-address-patterns`.
+
+## v0.4.24
+
+FEATURE: New `socks` backend mode for use with private sharing. Use `zrok share private --backend-mode socks` and then `zrok access private` that share from somewhere else... very lightweight VPN-like functionality (https://github.com/openziti/zrok/issues/558)
+
+FEATURE: New `zrok admin create account` command that allows populating accounts directly into the underlying controller database (https://github.com/openziti/zrok/issues/551)
+
+CHANGE: The `zrok test loopback public` utility to report non-`200` errors and also ensure that the listening side of the test is fully established before starting loopback testing.
+
+CHANGE: The OpenZiti SDK for golang (https://github.com/openziti/sdk-golang) has been updated to version `v0.22.28`
 
 ## v0.4.23